+++ app/upgrades.go
@@ -454,6 +454,8 @@ func (app *StrideApp) setupUpgradeHandlers(appOpts servertypes.AppOptions) {
 			app.configurator,
 			app.appCodec,
 			app.POAKeeper,
+			app.StakeibcKeeper,
+			app.InterchainqueryKeeper,
 		),
 	)